+++ ios/Classes/FacebookMessengerButtonProxy.m
@@ -1,0 +1,39 @@
+/**
+ * Appcelerator Titanium Mobile
+ * Copyright (c) 2009-2016 by Appcelerator, Inc. All Rights Reserved.
+ * Licensed under the terms of the Apache Public License
+ * Please see the LICENSE included with this distribution for details.
+ */
+
+#import "FacebookMessengerButtonProxy.h"
+#import "FacebookMessengerButton.h"
+#import "TiUtils.h"
+
+@implementation FacebookMessengerButtonProxy
+
+#pragma mark Layout Helper
+
+-(UIViewAutoresizing)verifyAutoresizing:(UIViewAutoresizing)suggestedResizing
+{
+    return suggestedResizing & ~(UIViewAutoresizingFlexibleHeight|UIViewAutoresizingFlexibleWidth);
+}
+
+-(void)viewDidAttach
+{
+    [(FacebookMessengerButton*)[self view] messengerButton];
+}
+
+USE_VIEW_FOR_VERIFY_HEIGHT
+USE_VIEW_FOR_VERIFY_WIDTH
+
+-(TiDimension)defaultAutoWidthBehavior:(id)unused
+{
+    return TiDimensionAutoSize;
+}
+
+-(TiDimension)defaultAutoHeightBehavior:(id)unused
+{
+    return TiDimensionAutoSize;
+}
+
+@end